--- a/test/fixtures/ES6/arrow-function/array-binding-pattern/invalid-dup-param.failure.json
+++ b/test/fixtures/ES6/arrow-function/array-binding-pattern/invalid-dup-param.failure.json
@@ -1,0 +1,1 @@
+{"index":14,"lineNumber":1,"column":15,"message":"Error: Line 1: Strict mode function may not have duplicate parameter names","description":"Strict mode function may not have duplicate parameter names"}