--- a/test/fixtures/ES6/arrow-function/array-binding-pattern/invalid-elision-after-rest.failure.json
+++ b/test/fixtures/ES6/arrow-function/array-binding-pattern/invalid-elision-after-rest.failure.json
@@ -1,0 +1,1 @@
+{"index":11,"lineNumber":1,"column":12,"message":"Error: Line 1: Unexpected token =>","description":"Unexpected token =>"}