--- a/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-member-expr.failure.json
+++ b/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-member-expr.failure.json
@@ -1,0 +1,1 @@
+{"index":10,"lineNumber":1,"column":11,"message":"Error: Line 1: Unexpected token =>","description":"Unexpected token =>"}