--- a/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-method-in-pattern.failure.json
+++ b/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-method-in-pattern.failure.json
@@ -1,0 +1,1 @@
+{"index":14,"lineNumber":1,"column":15,"message":"Error: Line 1: Unexpected token =>","description":"Unexpected token =>"}