--- a/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-nested-param.failure.json
+++ b/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-nested-param.failure.json
@@ -1,0 +1,1 @@
+{"index":50,"lineNumber":1,"column":51,"message":"Error: Line 1: Unexpected token =>","description":"Unexpected token =>"}