--- a/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-pattern-without-parenthesis.failure.json
+++ b/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-pattern-without-parenthesis.failure.json
@@ -1,0 +1,1 @@
+{"index":3,"lineNumber":1,"column":4,"message":"Error: Line 1: Unexpected token =>","description":"Unexpected token =>"}