--- a/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-rest-in-object-pattern.failure.json
+++ b/test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-rest-in-object-pattern.failure.json
@@ -1,0 +1,1 @@
+{"index":4,"lineNumber":1,"column":5,"message":"Error: Line 1: Unexpected token ...","description":"Unexpected token ..."}