--- a/test/fixtures/ES6/binding-pattern/array-pattern/with-default-catch-param.tree.json
+++ b/test/fixtures/ES6/binding-pattern/array-pattern/with-default-catch-param.tree.json
@@ -86,12 +86,13 @@
                         "elements": [
                             {
                                 "range": [
-                                    null,
+                                    16,
                                     21
                                 ],
                                 "loc": {
                                     "start": {
-                                        "column": null
+                                        "line": 1,
+                                        "column": 16
                                     },
                                     "end": {
                                         "line": 1,
@@ -194,12 +195,13 @@
                     "elements": [
                         {
                             "range": [
-                                null,
+                                16,
                                 21
                             ],
                             "loc": {
                                 "start": {
-                                    "column": null
+                                    "line": 1,
+                                    "column": 16
                                 },
                                 "end": {
                                     "line": 1,