--- a/test/fixtures/ES6/binding-pattern/array-pattern/with-default-fn.tree.json
+++ b/test/fixtures/ES6/binding-pattern/array-pattern/with-default-fn.tree.json
@@ -69,12 +69,13 @@
                     "elements": [
                         {
                             "range": [
-                                null,
+                                12,
                                 15
                             ],
                             "loc": {
                                 "start": {
-                                    "column": null
+                                    "line": 1,
+                                    "column": 12
                                 },
                                 "end": {
                                     "line": 1,