--- a/test/fixtures/ES6/destructuring-assignment/invalid-cover-grammar.failure.json
+++ b/test/fixtures/ES6/destructuring-assignment/invalid-cover-grammar.failure.json
@@ -1,0 +1,1 @@
+{"index":22,"lineNumber":1,"column":23,"message":"Error: Line 1: Unexpected token =","description":"Unexpected token ="}