--- a/test/fixtures/ES6/destructuring-assignment/invalid-group-assignment.failure.json
+++ b/test/fixtures/ES6/destructuring-assignment/invalid-group-assignment.failure.json
@@ -1,0 +1,1 @@
+{"index":5,"lineNumber":1,"column":6,"message":"Error: Line 1: Invalid left-hand side in assignment","description":"Invalid left-hand side in assignment"}