--- a/test/fixtures/ES6/destructuring-assignment/object-pattern/invalid-lhs-01.failure.json
+++ b/test/fixtures/ES6/destructuring-assignment/object-pattern/invalid-lhs-01.failure.json
@@ -1,0 +1,1 @@
+{"index":9,"lineNumber":1,"column":10,"message":"Error: Line 1: Invalid left-hand side in assignment","description":"Invalid left-hand side in assignment"}