--- a/test/fixtures/ES6/destructuring-assignment/object-pattern/invalid-lhs-02.failure.json
+++ b/test/fixtures/ES6/destructuring-assignment/object-pattern/invalid-lhs-02.failure.json
@@ -1,0 +1,1 @@
+{"index":10,"lineNumber":1,"column":11,"message":"Error: Line 1: Invalid left-hand side in assignment","description":"Invalid left-hand side in assignment"}