--- a/test/fixtures/ES6/destructuring-assignment/object-pattern/invalid-pattern-with-method.failure.json
+++ b/test/fixtures/ES6/destructuring-assignment/object-pattern/invalid-pattern-with-method.failure.json
@@ -1,0 +1,1 @@
+{"index":13,"lineNumber":1,"column":14,"message":"Error: Line 1: Invalid left-hand side in assignment","description":"Invalid left-hand side in assignment"}