--- a/test/fixtures/ES6/destructuring-assignment/object-pattern/object-pattern-assignment.js
+++ b/test/fixtures/ES6/destructuring-assignment/object-pattern/object-pattern-assignment.js
@@ -1,0 +1,8 @@
+({
+    a,
+    a:a,
+    a:a=a,
+    [a]:{a},
+    a:some_call()[a],
+    a:this.a
+} = 0);