… esprima.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/array-binding-pattern-01.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/array-binding-pattern-02.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/array-binding-pattern-03.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/array-binding-pattern-empty.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/elision.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/invalid-dup-param.js
… test/fixtures/ES6/arrow-function/array-binding-pattern/invalid-elision-after-rest.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-member-expr.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-method-in-pattern.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-nested-param.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-pattern-without-parenthesis.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/invalid-rest-in-object-pattern.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/nested-cover-grammar.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/nested-cover-grammar.tree.json
… test/fixtures/ES6/arrow-function/object-binding-pattern/object-binding-pattern-01.js
… test/fixtures/ES6/arrow-function/object-binding-pattern/object-binding-pattern-01.tree.json
… test/fixtures/ES6/arrow-function/object-binding-pattern/object-binding-pattern-empty.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/dup-assignment.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/dup-assignment.tree.json
… test/fixtures/ES6/destructuring-assignment/array-pattern/elision.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/member-expr-in-rest.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/member-expr-in-rest.tree.json
… test/fixtures/ES6/destructuring-assignment/array-pattern/nested-assignment.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/nested-assignment.tree.json
… test/fixtures/ES6/destructuring-assignment/array-pattern/nested-cover-grammar.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/nested-cover-grammar.tree.json
… test/fixtures/ES6/destructuring-assignment/array-pattern/simple-assignment.js
… test/fixtures/ES6/destructuring-assignment/array-pattern/simple-assignment.tree.json
… test/fixtures/ES6/destructuring-assignment/invalid-cover-grammar.js