Closes #1045.
Related to #1060.