@mikesherov 2015-03-18
Woah. Looks good so far from what I can tell. Some of the test files have a misspelling of `patter` instead of `pattern` and this doesn't actually close #1060 until we bring this back to the harmony branch.