@ikarienator 2015-03-18
I think it does close #1060 because the destructuring in for in loop will be supported in this PR as well.