@ikarienator 2015-03-18
@mikesherov I believe the attached test cases are actually more thorough than espree. @nzakas can you help review the tests?