@mikesherov 2015-03-19
@ikarienator LGTM but this can't close #1060 until all of the work is back ported to the harmony branch as well.

Also, please address the test files that have a misspelling of `patter` instead of `pattern`.

Great work here!