@ikarienator 2015-03-19
Amended it! Thanks guys for the review!