@ikarienator 2015-03-19
Before:

| Case | Size | Benchmark |
|:-----:|:----:|----------:|
|    Underscore 1.5.2 |   42.5 KiB  |   4.16 ms ± 1.10%  |
|      Backbone 1.1.0 |   58.7 KiB  |   4.20 ms ± 0.87%  |
|      MooTools 1.4.5 |  156.7 KiB  |  21.88 ms ± 0.81%  |
|        jQuery 1.9.1 |  262.1 KiB  |  21.80 ms ± 0.90%  |
|          YUI 3.12.0 |  330.4 KiB  |  23.16 ms ± 2.62%  |
| jQuery.Mobile 1.4.2 |  442.2 KiB  |  67.39 ms ± 2.57%  |
|       Angular 1.2.5 |  701.7 KiB  |  46.06 ms ± 2.63%  |
|             Summary | 1994.3 KiB  | 188.66 ms ± 2.26%  |

After:

| Case | Size | Benchmark |
|:-----:|:----:|----------:|
|    Underscore 1.5.2 |   42.5 KiB  |   4.94 ms ± 4.02%  |
|      Backbone 1.1.0 |   58.7 KiB  |   4.24 ms ± 0.76%  |
|      MooTools 1.4.5 |  156.7 KiB  |  22.24 ms ± 0.87%  |
|        jQuery 1.9.1 |  262.1 KiB  |  22.65 ms ± 0.77%  |
|          YUI 3.12.0 |  330.4 KiB  |  23.36 ms ± 2.98%  |
| jQuery.Mobile 1.4.2 |  442.2 KiB  |  62.07 ms ± 4.46%  |
|       Angular 1.2.5 |  701.7 KiB  |  48.55 ms ± 2.15%  |
|             Summary | 1994.3 KiB  | 188.04 ms ± 3.08%  |

It looks like that the performance impact is insignificant.