@ariya 2015-03-20
@ikarienator Nice benchmark numbers :smile: